move s3 to core
